frontend and backend implementation of comment, implemented AppService for general tasks, implemented snackbar instead of js alert

+++ frontend/src/app/components/bit-form/bit-form.component.ts
@@ -1,11 +1,11 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {CategoryService} from "../../services/category.service";
 import {BitService} from "../../services/bit.service";
 import {HttpClient} from "@angular/common/http";
-import {ActivatedRoute} from "@angular/router";
+import {ActivatedRoute, Router} from "@angular/router";
 import {FormControl, FormGroup, Validators} from "@angular/forms";
-import {Bit, Category} from "../../../types";
-import {Router} from "@angular/router";
+import {Category} from "../../../types";
+import {AppService} from "../../services/app.service";
 
 @Component({
   selector: 'app-bit-form',
@@ -17,7 +17,7 @@ export class BitFormComponent implements OnInit {
   bitFormGroup: FormGroup;
   categoryOptions: Category[] = [];
 
-  constructor(private http: HttpClient, private route: ActivatedRoute, private bitService: BitService, public categoryService: CategoryService, private router: Router) {
+  constructor(private http: HttpClient, private route: ActivatedRoute, private bitService: BitService, public categoryService: CategoryService, private router: Router, private AppService: AppService) {
     this.bitFormGroup = new FormGroup({
       id: new FormControl(null),
       title: new FormControl('', Validators.required),
@@ -41,22 +41,25 @@ export class BitFormComponent implements OnInit {
     const id = this.bitFormGroup.controls['id'].value
     if (id) {
       this.bitService.updateBit(this.bitFormGroup.value).subscribe(() => {
-        alert('Bit updated successfully!');
+        this.AppService.showSnackBar('Bit updated successfully!', 'hide');
       })
     } else {
       this.bitService.createBit(this.bitFormGroup.value).subscribe(() => {
-        alert('Bit created successfully!');
+        this.AppService.showSnackBar('Bit created successfully!', 'hide');
       })
     }
+    this.AppService.refreshRoute();
   }
 
   deleteBit() {
-    const id = this.bitFormGroup.controls['id'].value
-    if (id) {
-      this.bitService.deleteBit(this.bitFormGroup.value).subscribe(() => {
-        alert('Bit deleted successfully!');
-        this.router.navigate(['/'])
-      })
+    if (confirm("Are you sure to delete this bit?")) {
+      const id = this.bitFormGroup.controls['id'].value
+      if (id) {
+        this.bitService.deleteBit(this.bitFormGroup.value).subscribe(() => {
+          this.AppService.showSnackBar('Bit deleted successfully!', 'hide');
+          this.router.navigate(['/bitmap']);
+        })
+      }
     }
   }
